feature branch : move to using new locking level ints, and reinstate the runtime lock logging

+++ dso/branches/santiago_lock_rewrite/code/base/dso-l1-api/src/com/tc/object/bytecode/Manager.java
@@ -13,7 +13,7 @@ import com.tc.object.TCObject;
 import com.tc.object.event.DmiManager;
 import com.tc.object.loaders.ClassProvider;
 import com.tc.object.loaders.NamedClassLoader;
-import com.tc.object.lockmanager.api.LockLevel;
+import com.tc.object.locks.LockLevel;
 import com.tc.object.locks.TerracottaLocking;
 import com.tc.object.logging.InstrumentationLogger;
 import com.tc.properties.TCProperties;
@@ -33,10 +33,10 @@ public interface Manager extends TerracottaLocking {
   /** Bytecode type definition for this class */
   public static final String TYPE                        = "L" + CLASS + ";";
 
-  public final static int    LOCK_TYPE_READ              = LockLevel.READ;
-  public final static int    LOCK_TYPE_WRITE             = LockLevel.WRITE;
-  public final static int    LOCK_TYPE_CONCURRENT        = LockLevel.CONCURRENT;
-  public final static int    LOCK_TYPE_SYNCHRONOUS_WRITE = LockLevel.SYNCHRONOUS_WRITE;
+  public final static int    LOCK_TYPE_READ              = LockLevel.READ.toInt();
+  public final static int    LOCK_TYPE_WRITE             = LockLevel.WRITE.toInt();
+  public final static int    LOCK_TYPE_CONCURRENT        = LockLevel.CONCURRENT.toInt();
+  public final static int    LOCK_TYPE_SYNCHRONOUS_WRITE = LockLevel.SYNCHRONOUS_WRITE.toInt();
 
   /**
    * Determine whether this class is physically instrumented
